+++ railties/lib/rails/generators/rails/app/app_generator.rb
@@ -137,7 +137,9 @@ module Rails
       active_storage_config_exist    = File.exist?("config/storage.yml")
       rack_cors_config_exist         = File.exist?("config/initializers/cors.rb")
       assets_config_exist            = File.exist?("config/initializers/assets.rb")
-      csp_config_exist               = File.exist?("config/initializers/content_security_policy.rb")
+      asset_manifest_exist          = File.exist?("app/assets/config/manifest.js")
+      asset_app_stylesheet_exist    = File.exist?("app/assets/stylesheets/application.css")
+      csp_config_exist = File.exist?("config/initializers/content_security_policy.rb")
       permissions_policy_config_exist = File.exist?("config/initializers/permissions_policy.rb")
 
       @config_target_version = Rails.application.config.loaded_config_version || "5.0"
@@ -159,6 +161,14 @@ module Rails
 
       if options[:skip_sprockets] && !assets_config_exist
         remove_file "config/initializers/assets.rb"
+      end
+
+      if options[:skip_sprockets] && !asset_manifest_exist
+        remove_file "app/assets/config/manifest.js"
+      end
+
+      if options[:skip_sprockets] && !asset_app_stylesheet_exist
+        remove_file "app/assets/stylesheets/application.css"
       end
 
       unless rack_cors_config_exist
@@ -491,6 +501,8 @@ module Rails
       def delete_assets_initializer_skipping_sprockets
         if options[:skip_sprockets]
           remove_file "config/initializers/assets.rb"
+          remove_file "app/assets/config/manifest.js"
+          remove_file "app/assets/stylesheets/application.css"
         end
       end